+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V2.1.1 - 2020-7-5
+
+### Added
+
+- DEBUG: add additional debug output when successfully sending commands
+
+- DEBUG: dump out raw pool `controllerConfig` message from `node-screenlogic`
+
 ## V2.1.0 - 2020-7-4
 
 ### Added